refactor: split memory runtime from LLM tools
