+++ ai/__init__.py
@@ -1,0 +1,1 @@
+"""AI capability domains."""
