+++ ai/memory/__init__.py
@@ -1,0 +1,15 @@
+"""Long-term memory runtime and APIs."""
+
+from .service import (
+    check_mem0_status,
+    memory_forget,
+    memory_remember,
+    memory_search,
+)
+
+__all__ = [
+    "check_mem0_status",
+    "memory_forget",
+    "memory_remember",
+    "memory_search",
+]
